LayoutTests/fast/speech/change-focus-expected.txt

Tests for speech button click with <input type="text" speech>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Clicking on speech button.
PASS document.getElementById("input1").value is "Pictures of the moon"
PASS document.getElementById("input2").value is ""
PASS successfullyParsed is true

TEST COMPLETE
 
